% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAclasses.R
\docType{class}
\name{bayesmodelClass-class}
\alias{bayesmodelClass}
\alias{bayesmodelClass-class}
\title{S4 class for displaying output of Bayesian models}
\description{
S4 class for displaying the output of the functions \code{rrisk.BayesPEM} and
\code{rrisk.BayesZIP}.
}
\section{Slots}{

\describe{
\item{\code{convergence}}{{logical, whether the model has converged (assessed by the user only if GUI was used)}}

\item{\code{results}}{summary of runjags object}

\item{\code{jointposterior}}{{data frame giving the joint posterior probability distribution}}

\item{\code{nodes}}{{names of the parameters jointly estimated by the Bayes model}}

\item{\code{model}}{{model in rjags/JAGS syntax}}

\item{\code{chains}}{{number of independent MCMC chains}}

\item{\code{burn}}{{length of burn-in period}}

\item{\code{update}}{{length of update iterations for estimation}}
}}
\examples{
new("bayesmodelClass")
new("bayesmodelClass",model="Some model...", nodes="Nodes info...")
}

