% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{diagnostics}
\alias{diagnostics}
\title{Diagnostic plots for MCMC models provided by rrisk.BayesPEM, rrisk.BayesZIP and rrisk.BayesZINB functions}
\usage{
diagnostics(x, plots = FALSE)
}
\arguments{
\item{x}{Result slot returned by \code{\link{rrisk.BayesPEM}}, \code{\link{rrisk.BayesZIP}} or \code{\link{rrisk.BayesZINB}} of class \code{\link{runjags}}.}

\item{plots}{Logical, if \code{TRUE} the diagnostic plots will be displayed in separate windows}
}
\value{
Returns \code{TRUE} if the user confirms convergence. Otherwise the 
function returns \code{FALSE}.
}
\description{
This function provides a GUI for diagnostic plots to check convergence in Markov
chain Monte-Carlo (MCMC) models provided by 
\code{\link{rrisk.BayesZIP}}, \code{\link{rrisk.BayesZINB}} and \code{\link{rrisk.BayesPEM}}.
}
\details{
The argument \code{x} denotes the result of the MCMC models. The user is interactively requested to confirm whether 
the convergence has been reached. In this case the function returns 
\code{TRUE} otherwise \code{FALSE}.
\cr
The function is not intended to be called directly but is internally called
within the models.
}
\examples{
\dontrun{
resPEM <- rrisk.BayesPEM(x = 14, n = 100, k = 4, prior.se = c(64, 4), prior.sp = c(94, 16))
diagnostics(resPEM@results)
}
}
\keyword{manip}

