% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotDiag}
\alias{plotDiag}
\title{Auxiliary function}
\usage{
plotDiag(x, plotnumber)
}
\arguments{
\item{x}{runjags object}

\item{plotnumber}{single numerical value (1 for plotting Gelman-Rubin convergence
statistic (\code{samplesBgr()}), 2 for plotting history (\code{plotDensity}),
3 for plotting autocorrelation plots (\code{plotAutoC})) and 4 for plotting density plots}
}
\description{
Auxiliary function function for plotting graphical convergence
 diagnostics for Bayes models
}
\keyword{manip}

