% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAclasses.R
\docType{methods}
\name{show-methods}
\alias{show,bayesmodelClass-method}
\alias{show-methods}
\title{Show method for bayesmodelClass}
\usage{
\S4method{show}{bayesmodelClass}(object)
}
\arguments{
\item{object}{a \code{bayesmodelClass} object}
}
\description{
Show method for \code{\linkS4class{bayesmodelClass}}
}
\examples{
show(new("bayesmodelClass"))
}

