% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_rrisk.BayesZINB}
\alias{test_rrisk.BayesZINB}
\title{Quick test of the \code{rrisk.BayesZINB} function}
\usage{
test_rrisk.BayesZINB()
}
\value{
Returns an instance of the 
\code{\linkS4class{bayesmodelClass}} class.
}
\description{
The function is a quick test of the \code{\link{rrisk.BayesZINB}} function without setting parameters. Number of true negatives (n_true_neg) and true positives (n_true_pos) are sampled, all other parameters are fixed \cr
\code{prior.pi = c(1,1)} \cr
\code{lambda_true <- rgamma(n_true_pos, a=6, b=2)} \cr
\code{y_neg <- rep(0, n_true_neg)} \cr
\code{y_pos <- rpois(n_true_pos, lambda_true)} \cr
\code{y <- c(y_pos, y_neg)}
}
\examples{
\donttest{
------------------------------------------
test example
------------------------------------------

res <- test_rrisk.BayesZINB()
}
}

