% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_rrisk.BayesZIP}
\alias{test_rrisk.BayesZIP}
\title{Quick test of \code{rrisk.BayesZIP}}
\usage{
test_rrisk.BayesZIP()
}
\value{
Returns an instance of the 
\code{\linkS4class{bayesmodelClass}} class.
}
\description{
The function \code{\link{test_rrisk.BayesZIP}} is a quick test of the rrisk.BayesZIP function without setting parameters. Number of true negatives and true positives are sampled, all other parameters were fixed (\code{prior.lambda = c(0, 100)}, \code{prior.pi = c(1,1)}) .
}
\examples{
\donttest{
------------------------------------------
test example
------------------------------------------

res <- test_rrisk.BayesZIP()
}
}

