% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guis.R
\name{ZIPGUI}
\alias{ZIPGUI}
\title{Bayes estimation of a zero inflated Poisson (ZIP) model}
\usage{
ZIPGUI(data=NULL, prior.lambda=c(1,10), prior.pi=c(0.8,1),
 chains=3, burn=1000, update=10000, thin=1)
}
\arguments{
\item{data}{A vector of numeric data, containing zeros, and of minimal length 10.}

\item{prior.lambda}{Numeric vector containing minimum and maximum of a uniform
distribution used as prior for the Poisson parameter \code{lambda}, e.g., \cr \code{lambda} ~ \code{prior.lambda(*,*)=unif(*,*)}.}

\item{prior.pi}{Numeric vector containing parameters of a beta distribution
describing prior knowledge about prevalence (proportion of contaminated samples), e.g., \cr \code{pi} ~ \code{prior.pi(*,*)=beta(*,*)}.}

\item{chains}{Positive single numeric value, number of independent MCMC chains (default 3).}

\item{burn}{Positive single numeric value, length of the burn-in period (default 1000).}

\item{update}{Positive single numeric value, length of update iterations for estimation (default 10000).}

\item{thin}{Positive single numeric value (default 1). The samples from every kth iteration will be used for 
inference, where k is the value of thin. Setting \code{thin > 1} can help to reduce the autocorrelation
in the sample.}
}
\value{
The function \code{ZIPGUI} returns an instance of the \code{\linkS4class{bayesmodelClass}}
class containing following information
\item{\code{convergence}}{Logical, whether the model has converged (assessed by the user).}
\item{\code{results}}{Data frame containing statitsics of the posterior distribution.}
\item{\code{jointpost}}{Data frame giving the joint posterior probability distribution.}
\item{\code{nodes}}{Names of the parameters jointly estimated by the Bayes model.}
\item{\code{model}}{Model in BRugs/Winbugs syntax as a character string.}
\item{\code{chains}}{Number of independent MCMC chains.}
\item{\code{burn}}{Length of burn-in period.}
\item{\code{update}}{Length of update iterations for estimation.}
}
\description{
This function provides a GUI for the function \link{rrisk.BayesZIP}.
}
\note{
The convergence of the model will be entered by the user after the simulation process.
}
\examples{
\donttest{
data <- rpois(30, 4)
prior.lambda <- c(1, 10)
prior.pi <- c(0.8, 1)
ZIPGUI(data, prior.lambda, prior.pi)}
}
\references{
Bohning, D., E. Dietz, P. Schlattman, L. Mendonca, and U. Kirchner (1999). 
The zero-inflated Poisson model and the decayed, missing and filled teeth index in 
dental epidemiology. Journal of the Royal Statistical Society, Series A 162, 195-209.
}
\seealso{
\code{\link{rrisk.BayesZIP}}
}
\keyword{manip}

