% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAclasses.R
\docType{class}
\name{bayesmodelClass-class}
\alias{bayesmodelClass}
\alias{bayesmodelClass-class}
\title{S4 class for displaying output of Bayesian models}
\description{
S4 class for displaying the output of the functions \code{rrisk.BayesRGE} and
\code{rrisk.BayesZIP}.
}
\examples{
new("bayesmodelClass")
new("bayesmodelClass",model="Some model...", nodes="Nodes info...")
}

